move AffectedFilesCollector to ChangesReporting
