+++ packages/changes-reporting/src/ValueObject/RectorWithFileAndLineChange.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Rector\ChangesReporting\ValueObject;
 
+use Rector\Core\Contract\Rector\RectorInterface;
+use Rector\Core\Exception\NotRectorException;
+
 final class RectorWithFileAndLineChange
 {
     /**
@@ -23,6 +26,10 @@ final class RectorWithFileAndLineChange
 
     public function __construct(string $rectorClass, string $realPath, int $line)
     {
+        if (! is_a($rectorClass, RectorInterface::class, true)) {
+            throw new NotRectorException($rectorClass);
+        }
+
         $this->rectorClass = $rectorClass;
         $this->line = $line;
         $this->realPath = $realPath;
